+++ src/main/modules/http/http.ts
@@ -25,7 +25,7 @@ import {
 } from '../../../interface/http';
 import { Container } from '../../container';
 import { Module } from '..';
-import { logger } from '../../logger';
+import { logger } from '../../../shared/logger';
 
 export class HttpModule implements Module {
   readonly app: Express = express();